Add `import` alias
If stdout is a terminal, start Python with the given module imported.
The assumption is that I'm pasting in some Python code.

If stdout is not a Python terminal, complain and quit.
The obnoxiously generic-named `import` utility of ImageMagick is
available under /usr/bin/import, but chances are I'll never need to
use it.